Implement GET route in express and call inital GET in index.jsx
