+++ README.md
@@ -6,11 +6,11 @@
 - [ x ] Display a list of movies from data hardcoded in your index.jsx.
 - [ x ] Add a search bar so that a user may see if a movie is in the list. 
 - [ x ] After a user submits the search, display all matches (or partial matches) to that title.
-- [ ] Bonus: Handle the case of no movie by that name found gracefully.
+- [   ] Bonus: Handle the case of no movie by that name found gracefully.
 - [ x ] Add an input field for users to add movies.
-- [ ] Add a button to each list item that allows the user to toggle a 'watched' property.
-- [ ] Add two buttons to allow the users to toggle between a list of 'watched' movies and movies 'to watch'.
-- [ ] Add a panel of movie information that appears when the title is clicked (consider starting with hardcoded information)
+- [   ] Add a button to each list item that allows the user to toggle a 'watched' property.
+- [   ] Add two buttons to allow the users to toggle between a list of 'watched' movies and movies 'to watch'.
+- [   ] Add a panel of movie information that appears when the title is clicked (consider starting with hardcoded information)
 
 You can use the following data: 
 
@@ -32,9 +32,9 @@ var movies = [
 
 
 ## Level 2 - Express Server
-- [ ] Move the hardcoded data from index.jsx to your Express Server js file
-- [ ] In your Express Server code a GET route called /movies which will be used to get the hardcoded list of movies
-- [ ] Modify your index.jsx to call this /movies GET route to do an initial load of your data
+- [ x ] Move the hardcoded data from index.jsx to your Express Server js file
+- [ x ] In your Express Server code a GET route called /movies which will be used to get the hardcoded list of movies
+- [ x ] Modify your index.jsx to call this /movies GET route to do an initial load of your data
 - [ ] In your Express Server code a POST route called /movie. This route will allow the user to create a new movie
 - [ ] Modify your index.jsx use the /movie POST route
